requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
av=10.0.0=pypi_0
black=23.9.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.08.22=hecd8cb5_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.7=pypi_0
cloudpickle=2.2.1=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.11.0=pypi_0
detectron2=0.6=pypi_0
detectron2-densepose=0.6=pypi_0
ffmpeg=4.3=h0a44026_0
fonttools=4.42.1=pypi_0
freetype=2.12.1=hd8bbffd_0
fvcore=0.1.5.post20221221=pypi_0
gettext=0.21.0=he85b6c0_1
giflib=5.2.1=h6c40b1e_3
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
google-auth=2.23.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.58.0=pypi_0
hydra-core=1.3.2=pypi_0
icu=73.1=hcec6c5f_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
intel-openmp=2023.1.0=ha357a0b_43547
iopath=0.1.9=pypi_0
jpeg=9b=he5867d9_2
kiwisolver=1.4.5=pypi_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
libiconv=1.16=hca72f7f_2
libidn2=2.3.4=h6c40b1e_0
libpng=1.6.39=h6c40b1e_0
libtasn1=4.19.0=h6c40b1e_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.44.2=h6c40b1e_0
libwebp=1.2.0=hacca55c_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.10.4=h1bd7e62_1
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.3=pypi_0
mkl=2023.1.0=h8e150cf_43559
mkl-service=2.4.0=py38h6c40b1e_1
mkl_fft=1.3.8=py38h6c40b1e_0
mkl_random=1.2.4=py38ha357a0b_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=hcec6c5f_0
nettle=3.7.3=h230ac6f_1
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
numpy=1.23.5=py38h47b59a4_1
numpy-base=1.23.5=py38hcfaf2c3_1
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python-headless=4.8.0.76=pypi_0
openh264=2.1.1=h8346a28_0
openssl=3.0.10=hca72f7f_2
packaging=23.1=pypi_0
pathspec=0.11.2=pypi_0
pillow=9.3.0=py38h81888ad_1
pip=23.2.1=py38hecd8cb5_0
platformdirs=3.10.0=pypi_0
portalocker=2.8.2=pypi_0
protobuf=4.24.3=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.7=pypi_0
pyparsing=3.1.1=pypi_0
python=3.8.18=h5ee71fb_0
python-dateutil=2.8.2=pypi_0
pytorch=1.8.1=py3.8_0
pyyaml=6.0.1=pypi_0
readline=8.2=hca72f7f_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.0.0=py38hecd8cb5_0
six=1.16.0=pypi_0
sqlite=3.41.2=h6c40b1e_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=ha357a0b_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
termcolor=2.3.0=pypi_0
tk=8.6.12=h5d9f67b_0
tomli=2.0.1=pypi_0
torchaudio=0.8.1=py38
torchvision=0.9.1=py38_cpu
tqdm=4.66.1=pypi_0
typing_extensions=4.7.1=py38hecd8cb5_0
urllib3=1.26.16=pypi_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py38hecd8cb5_0
xz=5.4.2=h6c40b1e_0
yacs=0.1.8=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h4dc903c_0
zstd=1.4.9=h322a384_0
